SmartBookshelf/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pymupdf>=1.26.1",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
]
